backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pillow==11.1.0
opencv-python-headless==4.11.0.86
numpy==2.2.3
email-validator==2.2.0
psycopg[binary]==3.2.4
